Stream: implemented PROXY protocol V2 for the backend connections.
The PROXY protocol versions 1 and 2 were previously implemented for client
connections, but only version 1 was implemented for backend connections.

V2 can be configured using the "proxy_protocol_version <version>"
configuration directive.  Version 1 is used by default for compatibility
purposes.

TLVs for sending to backends can be configured using the
"proxy_protocol_tlv <name/hex_num> <value>" directive, where the first
argument is the name of the TLV type or the hexadecimal type number.  The
type name or number can be prepended with the "ssl_" prefix.  In this case,
the TLV will be sent as an SSL subtype.  The second argument of the
"proxy_protocol_tlv" directive is the TLV value.  It can contain variables.

A special "ssl_verify" type has been added to set the value of the "verify"
field in the PP2_TYPE_SSL TLV structure.  The PP2_CLIENT_SSL flag for the
"client" field of the PP2_TYPE_SSL structure is always set when any
SSL-related TLVs are used.  The PP2_CLIENT_CERT_SESS flag for the "client"
field is set when the PP2_SUBTYPE_SSL_CN TLV subtype is set.  The
PP2_CLIENT_CERT_CONN flag for the "client" field is also set if the client
connection uses TLS and does not reuse the TLS session.