HTTP/3: fixed client processing of retry packets.
According to RFC 9000,  17.2.5.2. Handling a Retry Packet:

    A client MUST accept and process at most one Retry packet for each
    connection attempt. After the client has received and processed an Initial
    or Retry packet from the server, it MUST discard any subsequent Retry
    packets that it receives.

The fix is to account the fact of processing a retry packet and ignore
any further retry packets.

The issue was triggered with openssl-3.5.0 which increased default size of
the ClientHello TLS message exceeding limit of 1200 bytes for Initial packets,
causing multiple Initial packets to be sent and multiple Retry packets received.