HTTP: allow multiple client blocks.
The change allows presence of multiple client blocks and changes the procedure
of binding a client block to the directive using it:

First, a directive tries to find a matching named location in any of client
blocks that present in configuration.

If such location is found, the directive will use configuration from this
location.

If such location is not found, a new (implicit) client block will be created
with default configuration.

Such approach allows to split global configurations needed by different
modules.

Note: to be found, client block must be declared prior to directives using them.

Example:

http {
    client {

        location @docker_events {
            ...
        }
        location @docker_containers {
            ...
        }
    }

    client {
        location @acme {
            ...
        }
    }
}