Upstream: answer with builtin error page instead last upstream page.
Now proxy answers with the standard error page in a case of receiving
errorous status from the upstream servers instead of retransmitting answer
from the last upstream.  There was a fuzzy logic.  The using of upstream
servers has no fixed sequence so the result can be different for the
several requests.

The upstream server status is fixed while receiving errorous answer when
the "proxy_next_upstream" directive is used with "5xx" codes.  These codes
can be considered as succesfull or errorous answer.  The upstream server
that has sent the last answer while "proxy_next_upstream" sequence had a
broken status.