Upstream: removed unneeded condition.
The removed condition allows to update "checked" time each time-out
interval.  The intermediate peer selections doesn't change this value.
Actually it doesn't make sense because the time-out is important in the
only case when the selected peer is failed but function is called for the
alive peers only.  So update "checked" time on each selection and optimize
out unneded condition.

Remove special case for the "single" mode due to the same code within
ngx_http_upstream_use_rr_peer() function.