HTTP/3: initialize "Host" header from authority.
Typically, HTTP/3 requests only contain ":authority" pseudo header
and not Host header.  This is a source of inconvenience for users
who are using $http_host variable.

The patch initializes value of the HTTP Host header from authority,
if the latter was missing.
This is a valid behavior according to RFC:

The RFC 9114 4.3.1 "Request Pseudo-Header Fields" says about :authority
and Host header field:

    If both fields are present, they MUST contain the same value.

This also makes behavior similar to the HTTP/2 processing of the
"authority" pseudo header.