ACME: removed requirement for "server" block to handle HTTP challenges.
This patch removes the need to define a separate "server" block with
a "listen 80" directive for ACME HTTP challenges. When HTTP challenge
handling is enabled, the ACME module now automatically opens a socket
on port 80.

The listening port can be customized using the new "acme_http_port"
directive, which has the same syntax as "acme_dns_port". While
customization is rarely needed, it can be useful for testing or
debugging.