Tests: relaxed slow_start requirements.
In the "p2 only" testcase we expect (ideally) all requests to be
routed to p2, since p2 is healthy, and p1 just started recovering
(i.e. has weight close to zero).

In reality, client spends time sending requests and we are not
accounting for it. A lot of time may pass between previous test
statement and running requests due to OS scheduling. Thus the race.

To mediate the problem, tolerate up to 10 requests being sent
to the 'wrong' backend and increase slow start time, thus increasing
smaller weight window.