QUIC: advanced bpf packet dispatching.
The current implementation has multiple issues in regard to reload and
binary upgrade processes:

 - the worker number is embedded into quic connection id what is insecure
 - the size of the socket map is fixed and defined by the first configuration
 - the upgrade does not work at all

The patch changes the process of accepting QUIC connections: instead
of a single socket used both for new and existing connections, two
sockets are used: one to listen for new connections, and another to
process existing connections.

With the help of the BPF-based dispatch, it is now possible to route
UDP packets to the proper sockets based on three tables that contain:

 - mapping of client connection ID to worker socket
    so if a packet belongs to an existing connection, it is passed to correct
    worker process immediately

 - mapping of worker number to listen socket
 - master state: number of workers and list of instances
   (the last map is global and not per-reuseport group as others)

So, the packet from new connection is dispatched as follows:

 1) select master instance; if there are multiple (i.e. 2) masters;
    the random master is selected, as both are functional.
    at this step we know how many workers in each master and obtain
    a pointer to map with listen sockets

 2) select one of the appropriate listening sockets using packet hash
    and a number of workers.

The bpf module updates corresponding maps when a new quic connection
is established or closed, and updates number of workers as reload/upgrade
happens.

On reload, listening sockets are closed and new connections are never
delivered to old workers. Existing connection continue to old workers.

The patch is based on Roman Arutyunyan <arut@nginx.com> original patchset.

Modifications (besides angie-specific) are fallback to non-bpf mode, when
it is not built or disabled. The original patch for the bpf module itself is
heavily reworked: the master indirection layer is added and the map creation
code is heavily refactored. The kernel bpf helper is modified accordingly.