ACME: added support for ALPN validation.
This patch adds support for the TLS ALPN validation method, which can
be enabled by specifying "alpn" in the "challenge" parameter of the
"acme_client" directive.

With this type of validation, the ACME server initiates a TLS
connection to the client on TCP port 443 and offers the
application-layer protocol "acme-tls/1" for negotiation. It also
includes an SNI extension containing the domain name being validated.
The ACME server then verifies that the "acme-tls/1" protocol was
successfully negotiated during the TLS handshake and that the
certificate presented contains extensions with expected values
transmitted to the client earlier according to the ACME protocol.
For more details, see RFC 8737 and RFC 8555.

The $acme_cert_* variables are used during verification to provide the
appropriate certificate and key to the ACME server.

A minimal working configuration might look like this:

    acme_client test https://localhost:14000/dir challenge=alpn ;

    server {
        listen               443 ssl;

        server_name          example.com;

        ssl_certificate      $acme_cert_test;
        ssl_certificate_key  $acme_cert_key_test;

        acme                 test;
    }