ACME: fixed ACME client disabling issue.
Previously, if an ACME client had been referenced in the "stream" block
but not in the "http" block, it was disabled during configuration
parsing, issued an "ACME client is not defined but referenced" message,
and never fetched a certificate. This has been fixed.

The patch also removes the requirement to place the "http" block
before the "stream" block.