ACME: added automatic calculation of certificate size.
Certificates obtained by ACME clients are stored in shared memory.
The size of a certificate depends significantly on the number of
domain names for which it is issued. Previously, the maximum
certificate size was specified by the "max_cert_size" parameter,
which defined the amount of shared memory to allocate. The default
value of this parameter was 8192.

Now, if the user does not explicitly assign a value to this
parameter, it is calculated automatically based on the number and
lengths of the domain names listed in the corresponding "server"
blocks. This ensures more efficient shared memory allocation
without user intervention.

A new directive, "acme_max_response_size", has also been added.
When many domain names are involved, responses from the ACME
server can become quite large. While the default size of the
response buffer is sufficient for most cases, there is a slight
chance that a response may exceed the buffer's capacity. In such
cases, certificate renewal will fail with the error: "Too big
subrequest response while sending to client".

To resolve this issue, users should increase the value of
"acme_max_response_size". However, it should be noted that this
will rarely be necessary.