ACME: fixed use of $acme_cert_* variables in "stream" block.
Previously, attempting to use a corresponding $acme_cert_* variable
in the "stream" block when no ACME client was referenced there using
"acme" directives would result in an "unknown variable" error.

Now, $acme_cert_* variables can be used in the "stream" block
regardless of whether the corresponding ACME clients are referenced
in that block. They must still be referenced in the "http" block.

The issue was introduced by ec906add4d33.