Core: avoid restoring shared memory addresss on configuration test.
If the "file=" option is specified in the "proxy_cache_path" directive,
and the cache state file is present, this leads to an attempt to restore
shared memory address during the configuration test.

The patch disables adress restoring during the configuration test.
The cache state header is still parsed and possible errors are reported.