ACME: resolved backward incompatiblity with "acme_http_port".
If there are binds on single IPs, binding on * is impossible on a Linux
kernel.  So, the previous approach to listen on *:80 by default prevented
the configuration from loading if there were only "listen IP:80" directives
in the previous configuration.

The new approach fixes this issue as well as prevents unintended listening on *.
Also, the directive name "acme_http_port" makes more sense now.

Now, if only port is specified (by default), the existing listeners on that port
are always used without an attempt to listen on *.

In the next major version, this change could be made to "acme_dns_port" as well.